Tower implemented
Arrows shoot out, but don't heat-seek. If we want that as a feature or upgrade I can add it.

I need to implement placement next. It will require a button on the HUD portion of the canvas and I'll have to figure out how to utilize the global Alpha value